Use Danger + GH Actions to handle labels on PRs
Initial stab at GitHub actions for some simple PR checks
